Conflicts:
	Release/CMakeLists.txt
	Release/src/CMakeLists.txt
	Release/tests/common/TestRunner/CMakeLists.txt

+++ Release/src/streams/fileio_posix.cpp
@@ -23,7 +23,7 @@
 * files. The supporting functions, which are in this file, use C-like signatures to avoid as many issues as
 * possible.
 *
-* For the latest on this and related APIs, please see http://casablanca.codeplex.com.
+* For the latest on this and related APIs, please see: https://github.com/Microsoft/cpprestsdk
 *
 * =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 ****/
@@ -560,7 +560,7 @@ size_t _seekrdtoend_fsb(Concurrency::streams::details::_file_info *info, int64_t
 
     if ( newpos == -1 ) return static_cast<size_t>(-1);
 
-    fInfo->m_rdpos = newpos / char_size;
+    fInfo->m_rdpos = static_cast<size_t> (newpos) / char_size;
     return fInfo->m_rdpos;
 }
